refactor(server): extract entity/geometry/ability classes from game.js
Split the 1500-line tail of game.js (geometry primitives, World, Player,
projectiles, hazards, aimers, abilities) into focused modules under
server/entities/, plus a neutral botEmotes.js helper so player.js can emit
bot emotes without a game<->player require cycle. game.js drops from 3930 to
2426 lines and now holds only the Room/Game/GameBoard orchestrators.

Pure mechanical move, no logic change; smoke test passes on all 50 maps.
