+++ server/game.js
@@ -8,6 +8,15 @@ var compressor = require('./compressor.js');
 var debug = require('./debug.js');
 var cellGraph = require('./cellGraph.js');
 var aiController = require('./aiController.js');
+var { emitBotEmote, botsCheerFor } = require('./botEmotes.js');
+var { Shape, Rect, Circle, Gate } = require('./entities/shapes.js');
+var { World } = require('./entities/world.js');
+var { Player, LobbyStartButton } = require('./entities/player.js');
+var { ExplosionAimer, SwapAimer } = require('./entities/aimers.js');
+var { Punch } = require('./entities/punch.js');
+var { HazardRail, Hazard, Bumper } = require('./entities/hazards.js');
+var { Projectile, CloudProj, SnowFlakeProj, BombProj, Puck } = require('./entities/projectiles.js');
+var { Ability, Blindfold, Swap, IceCannon, Bomb, SpeedBuff, SpeedDebuff, TileSwap, Cut, BombTrigger } = require('./entities/abilities.js');
 
 exports.getRoom = function (sig, size) {
 	return new Room(sig, size);
@@ -16,42 +25,6 @@ exports.getRoom = function (sig, size) {
 // Monotonic id source for headless AI racers. Bot ids are namespaced ("bot-N")
 // so they never collide with socket ids (which key human players in playerList).
 var botSeq = 0;
-
-// Reactive bot emotes: pick a chat-wheel emoji from the personality's emote set
-// for an event category (win/kill/hurt/cheer/greet) and broadcast a one-shot
-// botEmote. Each bot is rate-limited to one emote every 45-90s (randomized per bot
-// so they don't all chime at once and the chatter stays sparse). Bots use the full
-// emotional range — celebrating, clapping for others, reacting to hits — not just
-// taunting. The client renders it as a speech bubble. The cooldown lives on the
-// Player and is NOT cleared by reset(), so the cadence holds across rounds.
-function emitBotEmote(player, category) {
-	if (player == null || !player.isAI || player.profile == null) {
-		return;
-	}
-	var emotes = player.profile.emotes;
-	if (emotes == null || emotes[category] == null || emotes[category].length === 0) {
-		return;
-	}
-	var now = Date.now();
-	if (player.emoteReadyAt != null && now < player.emoteReadyAt) {
-		return;
-	}
-	player.emoteReadyAt = now + utils.getRandomInt(45000, 90000);
-	var opts = emotes[category];
-	var emote = opts[Math.floor(Math.random() * opts.length)];
-	messenger.messageRoomBySig(player.roomSig, "botEmote", { id: player.id, emote: emote });
-}
-
-// When a racer reaches the goal, other AI racers react — a wheel emote, often a
-// clap/acknowledgement (cheer), throttled so only a couple pipe up.
-function botsCheerFor(playerList, finisherId) {
-	for (var id in playerList) {
-		var p = playerList[id];
-		if (p.isAI && p.alive && id !== finisherId && Math.random() < 0.3) {
-			emitBotEmote(p, "cheer");
-		}
-	}
-}
 
 class Room {
 	constructor(sig, size) {
@@ -2450,1481 +2423,4 @@ class GameBoard {
 		return abilities;
 	}
 }
-
-class Shape {
-	constructor(x, y, color) {
-		this.x = x;
-		this.y = y;
-		this.newX = this.x;
-		this.newY = this.y;
-		this.velX = 0;
-		this.velY = 0;
-		this.color = color;
-	}
-	inBounds(shape) {
-		if (shape.radius) {
-			return this.testCircle(shape);
-		}
-		if (shape.width) {
-			return this.testRect(shape);
-		}
-		return false;
-	}
-}
-
-class Rect extends Shape {
-	constructor(x, y, width, height, angle, color) {
-		super(x, y, color);
-		this.width = width;
-		this.height = height;
-		this.angle = angle;
-		this.vertices = this.getVertices();
-	}
-	getVertices() {
-		var vertices = [];
-		var a = { x: this.x, y: this.y },
-			b = { x: this.width, y: this.y },
-			c = { x: this.width, y: this.height },
-			d = { x: this.x, y: this.height };
-
-		vertices.push(a, b, c, d);
-		return vertices;
-	}
-	pointInRect(objX, objY) {
-		var a = this.areaTriangle(this.vertices[0].x, this.vertices[0].y, this.vertices[1].x, this.vertices[1].y, this.vertices[2].x, this.vertices[2].y) +
-			this.areaTriangle(this.vertices[0].x, this.vertices[0].y, this.vertices[3].x, this.vertices[3].y, this.vertices[2].x, this.vertices[2].y);
-		var a1 = this.areaTriangle(objX, objY, this.vertices[0].x, this.vertices[0].y, this.vertices[1].x, this.vertices[1].y);
-		var a2 = this.areaTriangle(objX, objY, this.vertices[1].x, this.vertices[1].y, this.vertices[2].x, this.vertices[2].y);
-		var a3 = this.areaTriangle(objX, objY, this.vertices[2].x, this.vertices[2].y, this.vertices[3].x, this.vertices[3].y);
-		var a4 = this.areaTriangle(objX, objY, this.vertices[0].x, this.vertices[0].y, this.vertices[3].x, this.vertices[3].y);
-		return (a == a1 + a2 + a3 + a4);
-	}
-
-	areaTriangle(x1, y1, x2, y2, x3, y3) {
-		return Math.abs((x1 * (y2 - y3) + x2 * (y3 - y1) + x3 * (y1 - y2)) / 2.0);
-	}
-
-	getExtents() {
-		var minX = this.vertices[0].x,
-			maxX = minX,
-			minY = this.vertices[0].y,
-			maxY = minY;
-		for (var i = 0; i < this.vertices.length - 1; i++) {
-			var vert = this.vertices[i];
-			minX = (vert.x < minX) ? vert.x : minX;
-			maxX = (vert.x > maxX) ? vert.x : maxX;
-			minY = (vert.y < minY) ? vert.y : minY;
-			maxY = (vert.y > maxY) ? vert.y : maxY;
-		}
-		return { minX, maxX, minY, maxY };
-	}
-	testRect(rect) {
-		for (var i = 0; i < this.vertices.length; i++) {
-			if (rect.pointInRect(this.vertices[i].x, this.vertices[i].y)) {
-				return true;
-			}
-		}
-		for (var i = 0; i < rect.vertices.length; i++) {
-			if (this.pointInRect(rect.vertices[i].x, rect.vertices[i].y)) {
-				return true;
-			}
-		}
-		return false;
-	}
-	testCircle(circle) {
-		return circle.testRect(this);
-	}
-	getRandomLoc() {
-		return { x: Math.floor(Math.random() * (this.width - this.x)) + this.x, y: Math.floor(Math.random() * (this.height - this.y)) + this.y };
-	}
-	findFreeLoc(obj) {
-		var loc = this.getSafeLoc(obj.width || obj.radius);
-		return loc;
-	}
-	getSafeLoc(size) {
-		var objW = size + 5 + c.playerBaseRadius * 2;
-		var objH = size + 5 + c.playerBaseRadius * 2;
-		return { x: Math.floor(Math.random() * (this.width - 2 * objW - this.x)) + this.x + objW, y: Math.floor(Math.random() * (this.height - 2 * objH - this.y)) + this.y + objH, width: objW };
-	}
-}
-
-class Gate extends Rect {
-	constructor(x, y, width, height) {
-		super(x, y, width, height, 0, "grey");
-		this.isGate = true;
-	}
-	handleHit() {
-
-	}
-}
-
-class World extends Rect {
-	constructor(x, y, width, height, engine, playerList, hazardList, roomSig) {
-		super(x, y, width, height, 0, "white");
-		this.engine = engine;
-		this.playerList = playerList;
-		this.hazardList = hazardList;
-		this.roomSig = roomSig;
-		this.center = { x: width / 2, y: height / 2 };
-	}
-	update(dt) {
-
-	}
-	createNewPlayer(id) {
-		var color = this.getUniqueColorR();
-		var player = new Player(0, 0, 90, color, id, this.roomSig);
-		return player;
-	}
-	// A headless AI racer: a Player with no socket, flagged isAI, given an
-	// identity (name/title/personality) from the config cast. Reuses the same
-	// unique-color picker so bots are visually distinct from the human and each
-	// other. The bot brain (aiController) drives it via targetDirX/Y/braking.
-	createNewBot(id, identity) {
-		var color = this.getUniqueColorR();
-		var bot = new Player(0, 0, 90, color, id, this.roomSig);
-		bot.isAI = true;
-		bot.name = identity.name;
-		bot.title = identity.title;
-		bot.personality = identity.id;
-		// Full personality profile (trait weights) for the AI brain; the
-		// suffix-numbered duplicates share their base personality's traits.
-		bot.profile = identity;
-		return bot;
-	}
-	spawnPlayerRandomLoc(player) {
-		var loc = this.findFreeLoc(player);
-		player.initialLoc = loc;
-		player.x = loc.x;
-		player.y = loc.y;
-	}
-	setSpawnLocation(player) {
-		player.initialLoc = this.findFreeLoc(player);
-	}
-	getUniqueColorR() {
-		var usedColors = {};
-		for (var player in this.playerList) {
-			usedColors[this.playerList[player].color] = true;
-		}
-		return utils.getUniqueColor(usedColors);
-	}
-	resize() {
-		this.width = c.worldWidth;
-		this.height = c.worldHeight;
-		this.baseBoundRadius = this.width;
-		this.center = { x: this.width / 2, y: this.height / 2 };
-		this.engine.setWorldBounds(this.width, this.height);
-		var data = compressor.worldResize(this);
-		messenger.messageRoomBySig(this.roomSig, 'worldResize', data);
-	}
-}
-
-class Circle extends Shape {
-	constructor(x, y, radius, color) {
-		super(x, y, color);
-		this.radius = radius;
-	}
-	getExtents() {
-		return { minX: this.x - this.radius, maxX: this.x + this.radius, minY: this.y - this.radius, maxY: this.y + this.radius };
-	}
-
-	testCircle(circle) {
-		var objX1, objY1, objX2, objY2, distance;
-		objX1 = this.newX || this.x;
-		objY1 = this.newY || this.y;
-		objX2 = circle.newX || circle.x;
-		objY2 = circle.newY || circle.y;
-		distance = utils.getMag(objX2 - objX1, objY2 - objY1);
-		distance -= this.radius;
-		distance -= circle.radius;
-		if (distance <= 0) {
-			return true;
-		}
-		return false;
-	}
-
-	testRect(rect) {
-		if (this.lineIntersectCircle({ x: rect.x, y: rect.y }, { x: rect.newX, y: rect.newY })) {
-			return true;
-		}
-		if (rect.pointInRect(this.x, this.y)) {
-			return true;
-		}
-
-		if (this.lineIntersectCircle(rect.vertices[0], rect.vertices[1]) ||
-			this.lineIntersectCircle(rect.vertices[1], rect.vertices[2]) ||
-			this.lineIntersectCircle(rect.vertices[2], rect.vertices[3]) ||
-			this.lineIntersectCircle(rect.vertices[3], rect.vertices[0])) {
-			return true;
-		}
-
-		for (var i = 0; i < rect.vertices.length; i++) {
-			var distsq = utils.getMagSq(this.x, this.y, rect.vertices[i].x, rect.vertices[i].y);
-			if (distsq < Math.pow(this.radius, 2)) {
-				return true;
-			}
-		}
-		return false;
-	}
-	lineIntersectCircle(a, b) {
-		var ap, ab, dirAB, magAB, projMag, perp, perpMag;
-		ap = { x: this.x - a.x, y: this.y - a.y };
-		ab = { x: b.x - a.x, y: b.y - a.y };
-		magAB = Math.sqrt(utils.dotProduct(ab, ab));
-		dirAB = { x: ab.x / magAB, y: ab.y / magAB };
-
-		projMag = utils.dotProduct(ap, dirAB);
-
-		perp = { x: ap.x - projMag * dirAB.x, y: ap.y - projMag * dirAB.y };
-		perpMag = Math.sqrt(utils.dotProduct(perp, perp));
-		if ((0 < perpMag) && (perpMag < this.radius) && (0 < projMag) && (projMag < magAB)) {
-			return true;
-		}
-		return false;
-	}
-
-
-	getRandomCircleLoc(minR, maxR) {
-		var r = Math.floor(Math.random() * (maxR - minR));
-		var angle = Math.floor(Math.random() * (Math.PI * 2 - 0));
-		return { x: r * Math.cos(angle) + this.x, y: r * Math.sin(angle) + this.y };
-	}
-}
-
-class LobbyStartButton extends Circle {
-	constructor(x, y, angle, color) {
-		super(x, y, 75, color);
-		this.isLobbyStart = true;
-	}
-	handleHit(object) {
-
-	}
-}
-
-class Player extends Circle {
-	constructor(x, y, angle, color, id, roomSig) {
-		super(x, y, c.playerBaseRadius, color);
-		this.isPlayer = true;
-		this.enabled = true;
-		this.alive = true;
-		this.color = color;
-		this.id = id;
-		this.roomSig = roomSig;
-		this.currentState = null;
-		this.initialLoc = { x: 0, y: 0 };
-
-		//Sleep Variables
-		this.awake = true;
-		this.kick = false;
-		this.sleepWaitTime = c.playerStartSleepTime;
-		this.sleepTimer = null;
-		this.sleepTimeLeft = this.sleepWaitTime;
-
-		this.kickWaitTime = c.playerAFKKickTime;
-		this.kickTimer = null;
-		this.kickTimeLeft = this.kickWaitTime;
-
-		this.chatCoolDownWaitTime = 10;
-		this.chatCoolDownTimer = null;
-		this.chatCoolDownTimeLeft = this.chatCoolDownWaitTime;
-
-		//Game Variables
-		this.hittingLobbyButton = false;
-		this.reachedGoal = false;
-		this.timeReached = null;
-		this.collapseMargin = null; // distance to the lava front while collapsing (for clutch-goal cheer)
-		this.notches = 0;
-		this.nearVictory = false;
-		this.fellFromVictory = false;
-		this.infected = false;
-		this.isZombie = false;
-		this.exploded = false;
-
-		//Movement
-		this.moveForward = false;
-		this.moveBackward = false;
-		this.turnLeft = false;
-		this.turnRight = false;
-		this.attack = false;
-		this.angle = 0;
-
-
-		//Attack
-		this.acquiredAbility = null;
-		this.ability = null;
-		this.punch = null;
-		this.punchedBy = null;
-		this.murderedBy = null;
-
-		this.roundKills = 0;
-		this.multiKillCount = 0;
-		this.openMultiKillWindow = false;
-		this.killedPlayerList = [];
-
-
-		this.punchWaitTime = c.playerPunchCooldown;
-		this.punchedTimer = null;
-		this.punchTimeLeft = this.punchWaitTime;
-
-		//On Fire
-		this.onFire = 0;
-		this.fireTimer = null;
-		this.fireTimeLeft = 0;
-
-		//Lobby tutorial (no scoring). invulnUntil: timestamp grace window after a
-		//lobby respawn; onSanctuary: standing on a transparent "background" cell
-		//(neutral/immutable ground); lobbyRespawnPending: 'death' | 'goal' flagged by
-		//handleHit and consumed by GameBoard.updatePlayers (deferred like ability picks).
-		this.invulnUntil = 0;
-		this.invulnHeldInCircle = false;
-		this.onSanctuary = false;
-		this.lobbyRespawnPending = null;
-
-		//Achievements
-		this.savior = 0;
-		this.totalKills = 0;
-		this.brutalist = 0;
-		this.survivalist = 0;
-		this.resourceful = 0;
-		this.bully = 0;
-
-		//Engine Variables
-		this.newX = this.x;
-		this.newY = this.y;
-		this.velX = 0;
-		this.velY = 0;
-		this.dragMultiplier = 1;
-		this.dragCoeff = c.playerDragCoeff;
-		this.brakeCoeff = c.playerBrakeCoeff;
-		this.maxVelocity = c.playerMaxSpeed;
-		this.acel = c.playerBaseAcel;
-
-		this.currentSpeedBonus = 0;
-
-		//AI (set on bot players by world.createNewBot; humans leave these untouched)
-		this.isAI = false;
-		this.name = null;
-		this.title = null;
-		this.personality = null;
-		this.profile = null;
-		this.emoteReadyAt = 0;
-		//Steering outputs the engine's isAI branch reads each tick.
-		this.targetDirX = 0;
-		this.targetDirY = 0;
-		this.braking = false;
-		//Scratch space for the bot brain (aiController) — path cache, timers, etc.
-		this.ai = null;
-	}
-	update(currentState, dt, punchDirectional) {
-		this.currentState = currentState;
-		// Bots have no socket and never AFK — the sleep/kick path ends in
-		// messageClientBySig (Room.checkAFK), which would throw on a socketless
-		// bot. Skip it entirely; bots stay awake for their whole life.
-		if (!this.isAI) {
-			this.checkForSleep(currentState);
-		}
-		if (this.alive == false) {
-			return;
-		}
-		this.dt = dt;
-		this.move();
-		this.checkAttack(currentState, punchDirectional);
-		this.checkChatCoolDownTimer();
-		this.checkFireTimer();
-	}
-	move() {
-		this.x = this.newX;
-		this.y = this.newY;
-	}
-	checkAttack(currentState, punchDirectional) {
-		if (this.attack) {
-			if ((currentState == c.stateMap.racing || currentState == c.stateMap.collapsing || currentState == c.stateMap.lobby) && this.ability != null) {
-				this.punchedTimer = Date.now();
-				// No scoring in the lobby: firing an ability there must not tick the
-				// resourceful achievement stat (it isn't gated by checkForWinners). The
-				// curated lobby ability set is enforced by the map's tiles (bomb only),
-				// not here — swap/blindfold/tileSwap simply aren't placed to pick up.
-				if (currentState != c.stateMap.lobby) {
-					this.resourceful += 1;
-				}
-				this.ability.use();
-				// Clear the attack input so using an ability doesn't bleed into the
-				// next tick. Otherwise, once checkAbilities() nulls out the consumed
-				// ability, the still-true attack flag would trigger the punch
-				// slow-down in the engine (and throw a stray punch), stopping movement.
-				this.attack = false;
-				return;
-			}
-			// Punching is only meaningful during active play: the race itself
-			// (racing/collapsing), the pre-race gate where players jostle at the start
-			// (gated), and the lobby tutorial. In the non-play screens — the overview
-			// "next map" screen, waiting, and gameOver — swallow the input so no punch
-			// (and no punch sound, which the client only plays on the server's "punch"
-			// event) is thrown.
-			if (currentState != c.stateMap.racing && currentState != c.stateMap.collapsing && currentState != c.stateMap.lobby && currentState != c.stateMap.gated) {
-				this.attack = false;
-				return;
-			}
-			if (this.checkPunchCoolDown()) {
-				return;
-			}
-			this.punchedTimer = Date.now();
-			var punchRadius = c.punchRadius;
-			if (this.isZombie == true) {
-				punchRadius = c.brutalRounds.infection.punchRadius;
-			}
-			// Directional punch: place the hitbox in front of the player along its
-			// facing angle so you have to aim, and the radial knockback shoves the
-			// target forward. The decision (per round/mode) is resolved in
-			// GameBoard.updatePlayers and passed in. (Bumper/hazard punches are
-			// created elsewhere and stay centered/radial.)
-			var directional = punchDirectional === undefined ? c.directionalPunch : punchDirectional;
-			var punchX = this.x;
-			var punchY = this.y;
-			if (directional) {
-				var aim = utils.pos({ x: this.x, y: this.y }, c.punchReach, this.angle);
-				punchX = aim.x;
-				punchY = aim.y;
-			}
-			this.punch = new Punch(punchX, punchY, punchRadius, this.color, this.id, this.roomSig, 1, this.isZombie);
-			this.punch.directional = directional;
-			// No scoring in the lobby: punches still land (knockback feel) but must not
-			// tick the bully achievement stat, which isn't gated by checkForWinners.
-			if (currentState != c.stateMap.lobby) {
-				this.bully += 1;
-			}
-			messenger.messageRoomBySig(this.roomSig, "punch", compressor.sendPunch(this.punch));
-			this.attack = false;
-		}
-	}
-	checkPunchCoolDown() {
-		if (this.punchedTimer != null) {
-			this.punchTimeLeft = (this.punchWaitTime - (Date.now() - this.punchedTimer));
-			if (this.punchTimeLeft > 0) {
-				return true;
-			}
-			return false;
-		}
-	}
-	// Lobby-only protection.
-	// isTimedInvuln(): inside the timed post-respawn grace window.
-	// isInvuln(): timed grace OR "held" — parked in the start circle keeps it alive
-	//   (set by GameBoard.updateLobbyInvulnHold) so a griefed player can die, walk to
-	//   the center, and stand there safely until the game starts. During invuln,
-	//   lava/goal do nothing and knockback (cut/explosion/punch/bumper) is ignored.
-	// isProtected(): invuln OR standing on sanctuary (background) ground.
-	isTimedInvuln() {
-		return this.invulnUntil != 0 && Date.now() < this.invulnUntil;
-	}
-	isInvuln() {
-		return this.isTimedInvuln() || this.invulnHeldInCircle;
-	}
-	isProtected() {
-		return this.isInvuln() || this.onSanctuary;
-	}
-	checkChatCoolDownTimer() {
-		if (this.chatCoolDownTimer != null) {
-			this.chatCoolDownTimeLeft = ((this.chatCoolDownWaitTime * 1000 - (Date.now() - this.chatCoolDownTimer)) / (1000)).toFixed(1);
-			if (this.chatCoolDownTimeLeft > 0) {
-				return;
-			}
-			this.chatCoolDownTimer = null;
-		}
-	}
-	checkFireTimer() {
-		if (this.fireTimer != null) {
-			this.fireTimeLeft = ((this.onFire - (Date.now() - this.fireTimer)) / (1000)).toFixed(1);
-			messenger.messageRoomBySig(this.roomSig, "onFire", { owner: this.id, value: this.fireTimeLeft * 1000 });
-			if (this.fireTimeLeft > 0) {
-				return;
-			}
-			this.onFire = 0;
-			this.fireTimer = null;
-			messenger.messageRoomBySig(this.roomSig, "onFire", { owner: this.id, value: 0 });
-		}
-	}
-	setPunchedBy(owner) {
-		this.punchedBy = owner;
-		setTimeout(function (myself) {
-			if (myself.alive) {
-				myself.punchedBy = null;
-			}
-		}, c.playerKillWindow, this);
-	}
-	addKill(player) {
-		if (this.alive == false || this.isZombie == true) {
-			return;
-		}
-		this.killedPlayerList.push(player.id);
-		clearTimeout(multiKillIndex);
-		this.roundKills += 1;
-		this.totalKills += 1;
-		emitBotEmote(this, "kill");
-		this.addFire(c.playerKillFireBonus);
-		if (player.fellFromVictory) {
-			this.savior += 1;
-			this.addFire(c.playerKilledNearVictoryBonus);
-		}
-		if (this.openMultiKillWindow == true) {
-			if (this.multiKillCount == 0) {
-				this.multiKillCount = 2;
-			} else {
-				this.multiKillCount++;
-			}
-			this.addFire(c.playerMultiKillFireBonus);
-			messenger.messageRoomBySig(this.roomSig, "multiKill", this.multiKillCount);
-		}
-		if (this.totalKills == 5) {
-			messenger.messageRoomBySig(this.roomSig, "killingSpree", this.id);
-		}
-		if (this.totalKills == 10) {
-			messenger.messageRoomBySig(this.roomSig, "rampage", this.id);
-		}
-		if (this.totalKills == 15) {
-			messenger.messageRoomBySig(this.roomSig, "godLike", this.id);
-		}
-		var multiKillIndex = setTimeout(function (myself) {
-			myself.openMultiKillWindow = false;
-			myself.multiKillCount = 0;
-		}, c.playerMultiKillWindow, this)
-		this.openMultiKillWindow = true;
-	}
-	addFire(value) {
-		if (this.isZombie) {
-			return;
-		}
-		this.onFire += value;
-		messenger.messageRoomBySig(this.roomSig, "onFire", { owner: this.id, value: this.onFire });
-	}
-	addSpeed(newValue) {
-		//New speed cant go above max speed
-		var delta = 0;
-		if (newValue + this.currentSpeedBonus > c.playerMaxSpeedBonus) {
-			delta = c.playerMaxSpeedBonus - this.currentSpeedBonus;
-			this.currentSpeedBonus = c.playerMaxSpeedBonus;
-			return delta;
-		}
-		delta = newValue;
-		this.currentSpeedBonus += newValue;
-		return delta;
-	}
-	removeSpeed(newValue) {
-		//Subtract from Speedbonus cant go below 0
-		var delta = 0;
-		if (this.currentSpeedBonus - newValue < 0) {
-			delta = this.currentSpeedBonus;
-			this.currentSpeedBonus = 0;
-			return delta;
-		}
-		this.currentSpeedBonus -= newValue;
-		return newValue;
-	}
-	setSpeedBonus(newValue) {
-		if (newValue > c.playerMaxSpeedBonus) {
-			return;
-		}
-		this.currentSpeedBonus = newValue;
-	}
-	getSpeedBonus() {
-		return this.currentSpeedBonus;
-	}
-	increaseDragMultiplier(newValue) {
-		this.dragMultiplier *= newValue;
-	}
-	decreaseDragMultiplier(newValue) {
-		this.dragMultiplier = this.dragMultiplier / newValue;
-	}
-	getDragBonus() {
-		return this.dragMultiplier;
-	}
-	wakeUp() {
-		this.sleepTimer = null;
-		this.kickTimer = null;
-		if (this.awake == false) {
-			this.awake = true;
-			messenger.messageRoomBySig(this.roomSig, "playerAwake", this.id);
-		}
-	}
-	checkForSleep(currentState) {
-		if (this.sleepTimer != null) {
-			this.sleepTimeLeft = ((this.sleepWaitTime * 1000 - (Date.now() - this.sleepTimer)) / (1000)).toFixed(1);
-			if (this.sleepTimeLeft > 0) {
-				return;
-			}
-			this.checkAFK(currentState);
-			return;
-		}
-		this.sleepTimer = Date.now();
-	}
-	checkAFK(currentState) {
-		if (this.awake == true) {
-			if (currentState == c.stateMap.waiting || currentState == c.stateMap.lobby) {
-				debug.log("checkAFK: kicking player id=", this.id, " from state=", currentState, " (awake AFK)");
-				this.kick = true;
-				return;
-			}
-			this.awake = false;
-			messenger.messageRoomBySig(this.roomSig, "playerSleeping", this.id);
-		}
-		if (this.kickTimer != null) {
-			this.kickTimeLeft = ((this.kickWaitTime * 1000 - (Date.now() - this.kickTimer)) / (1000)).toFixed(1);
-			if (this.kickTimeLeft > 0) {
-				return;
-			}
-			debug.log("checkAFK: kicking player id=", this.id, " from state=", currentState, " (kickTimer expired)");
-			this.kick = true;
-			return;
-		}
-		this.kickTimer = Date.now();
-
-	}
-	resurrect(packet) {
-		if (packet.currentState != c.stateMap.racing &&
-			packet.currentState != c.stateMap.collapsing) {
-			return;
-		}
-		packet.isZombie = true;
-		packet.enabled = true;
-		packet.alive = true;
-		messenger.messageRoomBySig(packet.roomSig, "playerInfected", packet.id);
-	}
-	infect() {
-		if (this.infected == true) {
-			return;
-		}
-		this.infected = true;
-		var infectTimer = 0;
-		if (this.alive == true) {
-			infectTimer = utils.getRandomInt(1000, 4000);
-			setTimeout(this.killPlayer, infectTimer, this);
-		}
-		setTimeout(this.resurrect, infectTimer + 1500, this);
-	}
-	applyInfectedMods(object) {
-		this.acel = object.acel * c.brutalRounds.infection.acelModifer;
-		this.dragCoeff = object.dragCoeff * c.brutalRounds.infection.dragModifer;
-		this.brakeCoeff = object.brakeCoeff * c.brutalRounds.infection.brakeModifer;
-	}
-	// Restore normal ground grip (accel/drag/brake). Called when a player is off every
-	// map cell — e.g. shoved back behind the starting gate, a region with no terrain
-	// tile to stamp physics. Without this, the last tile they touched (ice in particular)
-	// keeps them sliding there until they step back onto real terrain. Zombies keep their
-	// infection handicap, measured against normal ground.
-	resetGrip() {
-		if (this.isZombie == true) {
-			this.acel = c.playerBaseAcel * c.brutalRounds.infection.acelModifer;
-			this.dragCoeff = c.playerDragCoeff * c.brutalRounds.infection.dragModifer;
-			this.brakeCoeff = c.playerBrakeCoeff * c.brutalRounds.infection.brakeModifer;
-			return;
-		}
-		this.acel = c.playerBaseAcel;
-		this.dragCoeff = c.playerDragCoeff;
-		this.brakeCoeff = c.playerBrakeCoeff;
-	}
-	handleHit(object) {
-		if (object.isLobbyStart) {
-			this.hittingLobbyButton = true;
-			return;
-		}
-		if (object.isPunch && object.ownerId != this.id) {
-			// Invulnerable lobby players (freshly respawned, or held safe in the start
-			// circle) can't be punched — so a griefer can't shove them into lava or out
-			// of the safe circle. Normal players still bump freely. No-op outside lobby.
-			if (this.isInvuln()) {
-				return;
-			}
-			if (object.ownerInfected) {
-				this.infect();
-			}
-			if (!object.mapOwned) {
-				this.setPunchedBy(object.ownerId);
-			}
-			_engine.punchPlayer(this, object);
-			messenger.messageRoomBySig(this.roomSig, "playerPunched", { owner: object.ownerId, victim: this.id, x: this.x, y: this.y });
-			emitBotEmote(this, "hurt"); // an AI racer reacts to getting knocked
-			return;
-		}
-		if (object.isPuck) {
-			if (this.isInvuln()) {
-				return;
-			}
-			_engine.puckPlayer(object, this);
-			messenger.messageRoomBySig(this.roomSig, "playerPunched", { owner: object.ownerId, victim: this.id, x: this.x, y: this.y });
-			return;
-		}
-		if (object.isGate) {
-			this.killSelf();
-			return;
-		}
-		if (object.isMapCell) {
-			// Track sanctuary ground (the transparent background type) for the
-			// force-shield in isProtected(). A player only "is on" the one cell this
-			// hit reports, so this reflects their current footing every tick.
-			this.onSanctuary = (object.id == c.tileMap.background.id);
-			if (object.id == c.tileMap.background.id) {
-				// Transparent lobby "background" / sanctuary ground: behaves as normal
-				// terrain. Applying normal grip here is what resets a player's physics
-				// when they step off an island onto background.
-				if (this.isZombie == true) {
-					this.applyInfectedMods(object);
-					return;
-				}
-				this.acel = object.acel;
-				this.brakeCoeff = object.brakeCoeff;
-				this.dragCoeff = object.dragCoeff;
-				return;
-			}
-			if (object.id == c.tileMap.normal.id) {
-				if (this.isZombie == true) {
-					this.applyInfectedMods(object);
-					return;
-				}
-				this.acel = object.acel;
-				this.brakeCoeff = object.brakeCoeff;
-				this.dragCoeff = object.dragCoeff;
-				return;
-			}
-			if (object.id == c.tileMap.slow.id) {
-				if (this.isZombie == true) {
-					this.applyInfectedMods(object);
-					return;
-				}
-				this.acel = object.acel;
-				this.dragCoeff = object.dragCoeff;
-				this.brakeCoeff = object.brakeCoeff;
-				return;
-			}
-			if (object.id == c.tileMap.fast.id) {
-				if (this.isZombie == true) {
-					this.applyInfectedMods(object);
-					return;
-				}
-				this.acel = object.acel;
-				this.dragCoeff = object.dragCoeff;
-				this.brakeCoeff = object.brakeCoeff;
-				return;
-			}
-			if (object.id == c.tileMap.lava.id) {
-				if (this.currentState == c.stateMap.lobby) {
-					// Lobby lava is a teaching prop: cosmetic death + safe respawn, never
-					// the real kill path (no killPlayer/removeNotch). Invuln players in
-					// their post-respawn grace window are immune.
-					if (this.isInvuln()) {
-						return;
-					}
-					this.lobbyRespawnPending = "death";
-					return;
-				}
-				if (this.isZombie == true) {
-					this.acel = object.acel;
-					this.dragCoeff = object.dragCoeff;
-					this.brakeCoeff = object.brakeCoeff;
-					return;
-				}
-				if (this.onFire > 0) {
-					if (this.fireTimer == null) {
-						this.fireTimer = Date.now();
-					}
-					this.checkFireTimer();
-					return;
-				}
-				this.killSelf();
-				return;
-			}
-			if (object.id == c.tileMap.ice.id) {
-				this.acel = object.acel;
-				this.brakeCoeff = object.brakeCoeff;
-				this.dragCoeff = object.dragCoeff;
-				return;
-			}
-			if (object.id == c.tileMap.goal.id) {
-				if (this.currentState == c.stateMap.lobby) {
-					// Lobby goal teaches the win condition: play the win cue + respawn,
-					// but never conclude/score (no reachedGoal/timeReached/playerConcluded/
-					// addNotch). Invuln players in their grace window are ignored.
-					if (this.isInvuln()) {
-						return;
-					}
-					this.lobbyRespawnPending = "goal";
-					return;
-				}
-				if (this.isZombie == true) {
-					this.acel = object.acel;
-					this.brakeCoeff = object.brakeCoeff;
-					this.dragCoeff = object.dragCoeff;
-					return;
-				}
-				this.alive = false;
-				this.reachedGoal = true;
-				this.timeReached = Date.now();
-				// "clutch" = beat the lava to the goal: only when the lava front was
-				// genuinely close (collapseMargin stamped each collapsing tick), so an
-				// uncontested slow-collapse stroll-in doesn't trigger the big eruption.
-				var clutch = this.collapseMargin != null && this.collapseMargin < c.audienceClutchMargin;
-				messenger.messageRoomBySig(this.roomSig, "playerConcluded", { id: this.id, clutch: clutch });
-				return;
-			}
-			if (object.id == c.tileMap.abilities.blindfold.id) {
-				if (this.ability != null || this.isZombie) {
-					return;
-				}
-				this.ability = new Blindfold(this.id, this.roomSig);
-				this.acquiredAbility = { mapID: object.voronoiId };
-				messenger.messageRoomBySig(this.roomSig, "abilityAcquired", { owner: this.id, ability: object.id, voronoiId: object.voronoiId });
-				return;
-			}
-			if (object.id == c.tileMap.abilities.swap.id) {
-				if (this.ability != null || this.isZombie) {
-					return;
-				}
-				this.ability = new Swap(this.id, this.roomSig);
-				this.acquiredAbility = { mapID: object.voronoiId };
-				messenger.messageRoomBySig(this.roomSig, "abilityAcquired", { owner: this.id, ability: object.id, voronoiId: object.voronoiId });
-				return;
-			}
-			if (object.id == c.tileMap.abilities.bomb.id) {
-				if (this.ability != null || this.isZombie) {
-					return;
-				}
-				this.ability = new Bomb(this.id, this.roomSig);
-				this.acquiredAbility = { mapID: object.voronoiId };
-				messenger.messageRoomBySig(this.roomSig, "abilityAcquired", { owner: this.id, ability: object.id, voronoiId: object.voronoiId });
-				return;
-			}
-			if (object.id == c.tileMap.abilities.speedBuff.id) {
-				if (this.ability != null || this.isZombie) {
-					return;
-				}
-				this.ability = new SpeedBuff(this.id, this.roomSig);
-				this.acquiredAbility = { mapID: object.voronoiId };
-				messenger.messageRoomBySig(this.roomSig, "abilityAcquired", { owner: this.id, ability: object.id, voronoiId: object.voronoiId });
-				return;
-			}
-			if (object.id == c.tileMap.abilities.speedDebuff.id) {
-				if (this.ability != null || this.isZombie) {
-					return;
-				}
-				this.ability = new SpeedDebuff(this.id, this.roomSig);
-				this.acquiredAbility = { mapID: object.voronoiId };
-				messenger.messageRoomBySig(this.roomSig, "abilityAcquired", { owner: this.id, ability: object.id, voronoiId: object.voronoiId });
-				return;
-			}
-			if (object.id == c.tileMap.abilities.tileSwap.id) {
-				if (this.ability != null || this.isZombie) {
-					return;
-				}
-				this.ability = new TileSwap(this.id, this.roomSig);
-				this.acquiredAbility = { mapID: object.voronoiId };
-				messenger.messageRoomBySig(this.roomSig, "abilityAcquired", { owner: this.id, ability: object.id, voronoiId: object.voronoiId });
-				return;
-			}
-			if (object.id == c.tileMap.abilities.iceCannon.id) {
-				if (this.ability != null || this.isZombie) {
-					return;
-				}
-				this.ability = new IceCannon(this.id, this.roomSig);
-				this.acquiredAbility = { mapID: object.voronoiId };
-				messenger.messageRoomBySig(this.roomSig, "abilityAcquired", { owner: this.id, ability: object.id, voronoiId: object.voronoiId });
-				return;
-			}
-			if (object.id == c.tileMap.abilities.cut.id) {
-				if (this.ability != null || this.isZombie) {
-					return;
-				}
-				this.ability = new Cut(this.id, this.roomSig);
-				this.acquiredAbility = { mapID: object.voronoiId };
-				messenger.messageRoomBySig(this.roomSig, "abilityAcquired", { owner: this.id, ability: object.id, voronoiId: object.voronoiId });
-				return;
-			}
-
-		}
-	}
-	addNotch(notchesToWin) {
-		if (this.notches + 1 >= notchesToWin) {
-			this.notches = notchesToWin;
-			this.nearVictory = true;
-			return;
-		}
-		this.notches += 1;
-	}
-	removeNotch() {
-		if (this.notches > 0) {
-			if (this.nearVictory == true) {
-				this.fellFromVictory = true;
-			}
-			this.nearVictory = false;
-			this.notches -= 1;
-		}
-	}
-	killPlayer(packet) {
-		if (packet.currentState != c.stateMap.racing &&
-			packet.currentState != c.stateMap.collapsing) {
-			return;
-		}
-		if (packet.alive == false) {
-			return;
-		}
-		if (packet.punchedBy != null) {
-			packet.murderedBy = packet.punchedBy;
-		}
-		packet.removeNotch();
-		packet.enabled = false;
-		packet.alive = false;
-		packet.ability = null;
-		packet.newX = packet.x;
-		packet.newY = packet.y;
-		packet.velX = 0;
-		packet.velY = 0;
-		packet.onFire = 0;
-		packet.moveForward = false;
-		packet.moveBackward = false;
-		packet.turnLeft = false;
-		packet.turnRight = false;
-		packet.attack = false;
-		// "killed" marks a player-caused death (punched/cut/blown into oblivion or
-		// the lava) versus an environmental/self death, so the crowd only cheers
-		// kills — not someone driving themselves into the lava.
-		messenger.messageRoomBySig(packet.roomSig, "playerDied", { id: packet.id, killed: packet.murderedBy != null });
-	}
-	killSelf() {
-		this.killPlayer(this);
-	}
-	reset(currentState) {
-		this.alive = true;
-		this.enabled = true;
-		this.infected = false;
-		this.isZombie = false;
-		this.exploded = false;
-		// Drop any stale AI path from the previous round so the bot re-paths fresh
-		// on the new map instead of steering toward old-map coordinates. Identity,
-		// personality knobs and the emote cooldown live on the Player/profile and
-		// are intentionally preserved.
-		if (this.ai != null) { this.ai.waypoints = null; this.ai.repathTimer = 0; }
-		this.x = this.initialLoc.x;
-		this.y = this.initialLoc.y;
-		this.newX = this.x;
-		this.newY = this.y;
-		this.velX = 0;
-		this.velY = 0;
-		this.dragMultiplier = 1;
-		this.dragCoeff = c.playerDragCoeff;
-		this.brakeCoeff = c.playerBrakeCoeff;
-		this.maxVelocity = c.playerMaxSpeed;
-		this.acel = c.playerBaseAcel;
-		this.currentSpeedBonus = 0;
-		this.moveForward = false;
-		this.moveBackward = false;
-		this.turnLeft = false;
-		this.turnRight = false;
-		this.attack = false;
-		this.reachedGoal = false;
-		this.timeReached = null;
-		this.collapseMargin = null;
-		this.punch = null;
-		this.punchedBy = null;
-		this.murderedBy = null;
-		this.roundKills = 0;
-		this.fellFromVictory = false;
-		this.openMultiKillWindow = false;
-		this.multiKillCount = 0;
-		this.acquiredAbility = null;
-		this.angle = 315;
-		// Clear lobby-only state on every race (re)start so a lobby respawn's invuln
-		// grace / sanctuary flag / pending-respawn can never bleed into a real round.
-		this.invulnUntil = 0;
-		this.invulnHeldInCircle = false;
-		this.onSanctuary = false;
-		this.lobbyRespawnPending = null;
-		if (currentState == c.stateMap.gameOver) {
-			this.survivalist = 0;
-			this.brutalist = 0;
-			this.resourceful = 0;
-			this.bully = 0;
-			this.ability = null;
-			this.notches = 0;
-			this.totalKills = 0;
-			this.onFire = 0;
-			this.savior = 0;
-			this.killedPlayerList = [];
-		}
-	}
-}
-
-class ExplosionAimer extends Circle {
-	constructor(x, y, radius, color, ownerId, roomSig) {
-		super(x, y, radius, color);
-		this.ownerId = ownerId;
-		this.roomSig = roomSig;
-		this.isExplosionAimer = true;
-		this.alive = true;
-		this.targetListAry = [];
-
-		this.explode = false;
-		this.explodeWaitTime = c.explosionWarnTime;
-		this.explodeTimer = null;
-		this.explodeTimeLeft = this.explodeWaitTime;
-	}
-	update() {
-		if (!this.alive) {
-			return;
-		}
-		if (this.radius < c.explosionRadius) {
-			this.grow();
-		}
-		this.checkExplodeTimer();
-
-	}
-	grow() {
-		this.radius += 2;
-	}
-	checkExplodeTimer() {
-		if (this.explodeTimer != null) {
-			this.explodeTimeLeft = ((this.explodeWaitTime * 1000 - (Date.now() - this.explodeTimer)) / (1000)).toFixed(1);
-			if (this.explodeTimeLeft > 0) {
-				return;
-			}
-			this.explodeTimer = null;
-			this.explode = true;
-			messenger.messageRoomBySig(this.roomSig, "terminateAimer", this.ownerId);
-			return;
-		}
-		this.explodeTimer = Date.now();
-	}
-	killSelf() {
-		this.alive = false;
-	}
-	handleHit() {
-
-	}
-}
-
-class SwapAimer extends Circle {
-	constructor(x, y, radius, color, ownerId, roomSig) {
-		super(x, y, radius, color);
-		this.ownerId = ownerId;
-		this.roomSig = roomSig;
-		this.isSwapAimer = true;
-		this.alive = true;
-		this.targetList = {};
-		this.targetListAry = [];
-		var index = setInterval(function (aimer) {
-			aimer.targetList = {};
-			aimer.targetListAry = [];
-		}, 300, this);
-		this.index = index;
-	}
-	update(owner) {
-		if (owner.alive == false || owner.isZombie) {
-			this.alive = false;
-			messenger.messageRoomBySig(this.roomSig, "terminateAimer", this.ownerId);
-			return;
-		}
-		if (this.radius < c.tileMap.abilities.swap.endSize) {
-			this.grow();
-		}
-		this.move(owner);
-	}
-	grow() {
-		this.radius += 2;
-	}
-	move(owner) {
-		this.x = owner.x;
-		this.y = owner.y;
-	}
-	handleHit(object) {
-		if (object.isPlayer && object.id != this.ownerId && object.isZombie == false && this.alive) {
-			if (this.targetList[object.id] == null) {
-				this.targetList[object.id] = object;
-				this.targetListAry.push(object.id);
-			}
-			return;
-		}
-	}
-}
-
-class Punch extends Circle {
-	constructor(x, y, radius, color, ownerId, roomSig, punchBonus, infected) {
-		super(x, y, radius, color);
-		this.ownerId = ownerId;
-		this.roomSig = roomSig;
-		this.isPunch = true;
-		this.ownerInfected = infected;
-		this.punchBonus = punchBonus;
-		this.type = "player";
-		this.mapOwned = false;
-	}
-	handleHit(object) {
-
-	}
-	getBonus() {
-		return this.punchBonus;
-	}
-}
-
-class HazardRail extends Rect {
-	constructor(x, y, width, height, angle, color, ownerId, roomSig) {
-		super(x, y, width, height, angle, color);
-		this.alive = true;
-		this.ownerId = ownerId;
-		this.roomSig = roomSig;
-		this.lengthSq = this.width * this.width;
-	}
-	update() {
-		if (this.alive == false) {
-			return;
-		}
-	}
-	handleHit(object) {
-		console.log("hazard rail hit");
-	}
-}
-
-class Hazard extends Circle {
-	constructor(x, y, radius, color, ownerId, roomSig, rail) {
-		super(x, y, radius, color);
-		this.alive = true;
-		this.ownerId = ownerId;
-		this.roomSig = roomSig;
-		this.moveable = false;
-		this.id = -1;
-		this.speed = 0;
-		this.angle = 0;
-		this.lengthSq = this.radius * radius;
-		if (rail != null) {
-			this.moveable = true;
-			this.rail = rail;
-		}
-	}
-	update() {
-		if (this.alive == false) {
-			return;
-		}
-		this.move();
-	}
-	move() {
-		this.x = this.newX;
-		this.y = this.newY;
-	}
-	handleHit(object) {
-
-	}
-}
-class Bumper extends Hazard {
-	constructor(x, y, radius, color, ownerId, roomSig, rail) {
-		super(x, y, radius, color, ownerId, roomSig, rail);
-		this.id = c.hazards.bumper.id;
-		if (this.rail != null) {
-			this.speed = c.hazards.movingBumper.speed;
-			this.id = c.hazards.movingBumper.id;
-			this.angle = this.rail.angle;
-		}
-		this.punch = null;
-	}
-	handleHit(object) {
-		if (!object.isPlayer && !object.isPuck) {
-			return;
-		}
-		if (this.punch == null) {
-			this.punch = new Punch(this.x, this.y, c.hazards.bumper.attackRadius, c.hazards.bumper.color, this.ownerId, this.roomSig, c.hazards.bumper.punchBonus, false);
-			this.punch.mapOwned = true;
-			this.punch.type = "bumper";
-		}
-	}
-}
-
-class Projectile extends Circle {
-	constructor(x, y, radius, color, ownerId, roomSig, angle) {
-		super(x, y, radius, color);
-		this.alive = true;
-		this.bounced = false;
-		this.ownerId = ownerId;
-		this.roomSig = roomSig;
-		this.angle = angle;
-		this.speed = 0;
-		this.velX = 0;
-		this.velY = 0;
-		this.newX = this.x;
-		this.newY = this.y;
-		this.type = "";
-	}
-	update() {
-		if (!this.alive) {
-			return;
-		}
-		this.checkForBounce();
-		this.move();
-	}
-	move() {
-		this.x = this.newX;
-		this.y = this.newY;
-	}
-	checkForBounce() {
-		if (this.bounced) {
-			this.bounced = false;
-			messenger.messageRoomBySig(this.roomSig, "projBounced");
-		}
-	}
-	handleHit(object) {
-
-	}
-}
-class CloudProj extends Projectile {
-	constructor(x, y, radius, color, ownerId, roomSig, angle) {
-		super(x, y, radius, color, ownerId, roomSig, angle);
-		this.speed = c.brutalRounds.cloudy.speed;
-		this.type = "cloud";
-	}
-	update() {
-		if (!this.alive) {
-			return;
-		}
-		this.move();
-	}
-}
-class SnowFlakeProj extends Projectile {
-	constructor(x, y, radius, color, ownerId, roomSig, angle) {
-		super(x, y, radius, color, ownerId, roomSig, angle);
-		this.explosionRadius = c.tileMap.abilities.iceCannon.explosionRadius;
-		this.speed = c.tileMap.abilities.iceCannon.speed;
-		this.explodeIce = false;
-		this.type = "snowFlake";
-		this.tileChanges = {};
-
-		this.explodeWaitTime = c.tileMap.abilities.iceCannon.lifetime;
-		this.explodeTimer = null;
-		this.explodeTimeLeft = this.explodeWaitTime;
-	}
-	update() {
-		this.checkExplodeTimer();
-		super.update();
-	}
-	checkExplodeTimer() {
-		if (this.explodeTimer != null) {
-			this.explodeTimeLeft = ((this.explodeWaitTime * 1000 - (Date.now() - this.explodeTimer)) / (1000)).toFixed(1);
-			if (this.explodeTimeLeft > 0) {
-				return;
-			}
-			this.explodeFlake();
-			return;
-		}
-		this.explodeTimer = Date.now();
-	}
-	explodeFlake() {
-		this.explodeIce = true;
-		this.alive = false;
-	}
-	handleHit(object) {
-		if (object.isMapCell) {
-			if (object.id != c.tileMap.lava.id && object.id != c.tileMap.goal.id && object.id != c.tileMap.slow.id) {
-				this.tileChanges[object.voronoiId] = c.tileMap.ice.id;
-			}
-		}
-	}
-}
-class BombProj extends Projectile {
-	constructor(x, y, radius, color, ownerId, roomSig, angle) {
-		super(x, y, radius, color, ownerId, roomSig, angle);
-		this.explosionRadius = c.tileMap.abilities.bomb.explosionRadius;
-		this.speed = c.tileMap.abilities.bomb.speed;
-		this.explode = false;
-		this.type = "bomb";
-
-		this.explodeWaitTime = c.tileMap.abilities.bomb.lifetime;
-		this.explodeTimer = null;
-		this.explodeTimeLeft = this.explodeWaitTime;
-	}
-	update() {
-		this.checkExplodeTimer();
-		super.update();
-	}
-	checkExplodeTimer() {
-		if (this.explodeTimer != null) {
-			this.explodeTimeLeft = ((this.explodeWaitTime * 1000 - (Date.now() - this.explodeTimer)) / (1000)).toFixed(1);
-			if (this.explodeTimeLeft > 0) {
-				return;
-			}
-			this.explodeBomb();
-			return;
-		}
-		this.explodeTimer = Date.now();
-	}
-	explodeBomb() {
-		this.explode = true;
-		this.alive = false;
-	}
-}
-
-class Puck extends Projectile {
-	constructor(x, y, radius, color, ownerId, roomSig, angle) {
-		super(x, y, radius, color, ownerId, roomSig, angle);
-		this.type = "puck";
-		this.isPuck = true;
-		this.speed = c.brutalRounds.hockey.baseSpeed;
-	}
-	handleHit(object) {
-		if (object.isPunch && object.ownerId != this.ownerId) {
-			_engine.punchPuck(this, object);
-			messenger.messageRoomBySig(this.roomSig, "playerPunched", { owner: object.ownerId, victim: this.id, x: this.x, y: this.y });
-			return;
-		}
-	}
-}
-
-class Ability {
-	constructor(owner, roomSig) {
-		this.id = null;
-		this.roomSig = roomSig;
-		this.ownerId = owner;
-		this.alive = true;
-	}
-	update() {
-
-	}
-	use() {
-		console.log("unimplemented");
-	}
-}
-class Blindfold extends Ability {
-	constructor(owner, roomSig) {
-		super(owner, roomSig);
-		this.id = c.tileMap.abilities.blindfold.id;
-		this.blind = false;
-	}
-	use() {
-		if (this.alive == false) {
-			return;
-		}
-		this.alive = false;
-		// Flag the room-wide blind so checkAbilities can record it for the AI
-		// vision-fairness self-handicap (bots read state, not pixels).
-		this.blind = true;
-		messenger.messageRoomBySig(this.roomSig, "blindfoldUsed", this.ownerId);
-	}
-}
-class Swap extends Ability {
-	constructor(owner, roomSig) {
-		super(owner, roomSig);
-		this.swap = false;
-		this.id = c.tileMap.abilities.swap.id;
-	}
-	use() {
-		if (this.alive == false) {
-			return;
-		}
-		this.swap = true;
-		this.alive = false;
-		messenger.messageRoomBySig(this.roomSig, "swapUsed", this.ownerId);
-	}
-}
-class IceCannon extends Ability {
-	constructor(owner, roomSig) {
-		super(owner, roomSig);
-		this.spawnSnowFlake = false;
-		this.id = c.tileMap.abilities.iceCannon.id;
-	}
-	use() {
-		if (this.alive == false) {
-			return;
-		}
-		this.spawnSnowFlake = true;
-		this.alive = false;
-		messenger.messageRoomBySig(this.roomSig, "iceCannon", this.ownerId);
-	}
-}
-
-class Bomb extends Ability {
-	constructor(owner, roomSig) {
-		super(owner, roomSig);
-		this.spawnBomb = false;
-		this.id = c.tileMap.abilities.bomb.id;
-	}
-	use() {
-		if (this.alive == false) {
-			return;
-		}
-		this.spawnBomb = true;
-		this.alive = false;
-		messenger.messageRoomBySig(this.roomSig, "bombUsed", this.ownerId);
-	}
-}
-class SpeedBuff extends Ability {
-	constructor(owner, roomSig) {
-		super(owner, roomSig);
-		this.applyBuff = false;
-		this.id = c.tileMap.abilities.speedBuff.id;
-	}
-	use() {
-		if (this.alive == false) {
-			return;
-		}
-		this.alive = false;
-		this.applyBuff = true;
-		messenger.messageRoomBySig(this.roomSig, "speedBuff", this.ownerId);
-	}
-}
-
-class SpeedDebuff extends Ability {
-	constructor(owner, roomSig) {
-		super(owner, roomSig);
-		this.applyDebuff = false;
-		this.id = c.tileMap.abilities.speedDebuff.id;
-	}
-	use() {
-		if (this.alive == false) {
-			return;
-		}
-		this.alive = false;
-		this.applyDebuff = true;
-		messenger.messageRoomBySig(this.roomSig, "speedDebuff", this.ownerId);
-	}
-}
-class TileSwap extends Ability {
-	constructor(owner, roomSig) {
-		super(owner, roomSig);
-		this.tileSwap = false;
-		this.id = c.tileMap.abilities.tileSwap.id;
-	}
-	use() {
-		if (this.alive == false) {
-			return;
-		}
-		this.alive = false;
-		this.tileSwap = true;
-		messenger.messageRoomBySig(this.roomSig, "tileSwap", this.ownerId);
-	}
-}
-
-class Cut extends Ability {
-	constructor(owner, roomSig) {
-		super(owner, roomSig);
-		this.applyCut = false;
-		this.id = c.tileMap.abilities.cut.id;
-	}
-	use() {
-		if (this.alive == false) {
-			return;
-		}
-		this.alive = false;
-		this.applyCut = true;
-		messenger.messageRoomBySig(this.roomSig, "cutUsed", this.ownerId);
-	}
-}
-
-
-class BombTrigger extends Ability {
-	constructor(owner, roomSig) {
-		super(owner, roomSig);
-		this.explodeBomb = false;
-		this.id = c.tileMap.abilities.bombTrigger.id;
-	}
-	use() {
-		if (this.alive == false) {
-			return;
-		}
-		this.explodeBomb = true;
-		this.alive = false;
-		messenger.messageRoomBySig(this.roomSig, "bombTriggered", this.ownerId);
-	}
-}
-
-
